+++ app/src/main/java/com/github/theholydevil/storybook/ChapterReader.java
@@ -1,16 +1,6 @@
 package com.github.theholydevil.storybook;
 
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
 
 /**
  * Created by Stefan on 06.09.2015.
@@ -19,14 +9,13 @@ import android.widget.ImageView;
 public class ChapterReader {
     private ViewPager viewPager;
     private Book book;
-    FragmentAdapter fragmentAdapter;
 
-    public ChapterReader(FragmentManager fm, ViewPager pager, Book book)
+    public ChapterReader(ViewPager pager, Book book)
     {
         this.book = book;
-        fragmentAdapter = new FragmentAdapter(fm, this.book);
         this.viewPager = pager;
-        this.viewPager.setAdapter(fragmentAdapter);
+        BookHandler.getChapterReaderFragmentAdapter().setBook(this.book);
+        this.viewPager.setAdapter(BookHandler.getChapterReaderFragmentAdapter());
         //this.viewPager.setOffscreenPageLimit(3); //Didn't make anything more smooth, only increases memory usage, try something else
     }
 
@@ -48,75 +37,5 @@ public class ChapterReader {
         this.viewPager.setCurrentItem(book.getChapterPageIndex(chapterIndex));
     }
 
-    public static class FragmentAdapter extends FragmentStatePagerAdapter
-    {
-        int count;
-        Book book;
 
-        public FragmentAdapter(FragmentManager fm, Book items)
-        {
-            super(fm);
-            this.book = items;
-            this.count = this.book.getPagePathList().size();
-        }
-
-        public Book getBook() {
-            return book;
-        }
-
-        @Override
-        public int getCount()
-        {
-            return count;
-        }
-
-        @Override
-        public Fragment getItem(int position){
-            return PageFragment.newInstance(position, this.book);
-        }
-    }
-
-    public static class PageFragment extends Fragment
-    {
-        //int pageIndex;
-        String path;
-
-        static PageFragment newInstance(int index, Book book)
-        {
-            PageFragment f = new PageFragment();
-
-            Bundle args = new Bundle();
-            //args.putInt("page", index);
-            args.putString("path", book.getItemPath(index));
-            f.setArguments(args);
-
-            return f;
-        }
-
-        @Override
-        public void onCreate(Bundle savedInstanceState)
-        {
-            super.onCreate(savedInstanceState);
-            //this.pageIndex = getArguments().getInt("page");
-            this.path = getArguments().getString("path");
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
-        {
-            View v = inflater.inflate(R.layout.fragment_page, container, false);
-            ImageView imageView = (ImageView) v.findViewById(R.id.imageView);
-
-            Bitmap bitmap = BitmapFactory.decodeFile(this.path);
-            imageView.setImageBitmap(bitmap);
-
-            return v;
-        }
-
-        @Override
-        public void onActivityCreated(Bundle savedInstanceState)
-        {
-            super.onActivityCreated(savedInstanceState);
-        }
-    }
 }
